# openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Statreg API
  version: 1.1.0
  description: API for managing publications, statistics and metadata for Statistics Norway
servers:
  - url: https://ssb.no/statreg-api
  - url: https://localhost:8080
...
components:
  schemas:
...
    Release:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        publish_time: { type: string, format: date-time }
        period_to: { type: string, format: date-time }
        period_from: { type: string, format: date-time }
...
    Release_update:
      allOf:
        - $ref: '#/components/schemas/Release'
        - type: object
          properties:
            comment: { type: string }
            release_date_precision: { type: string }
    Release_get:
      allOf:
        - $ref: '#/components/schemas/Release'
        - type: object
          properties:
            approval_status: { type: [string, 'null'] }
            statistic:
              type: object
              properties:
                shortname: { type: string, readOnly: true }
                name: { $ref: '#/components/schemas/Translations', readOnly: true }
    Release_details:
      allOf:
        - $ref: '#/components/schemas/Release_get'
        - type: object
          properties:
            has_versions: { type: boolean }
            variant:
              type: object
              properties:
                id: { type: integer, readOnly: true }
                frequency:
                  type: object
                  properties:
                    name: { $ref: '#/components/schemas/Translations', readOnly: true }
                revision:
                  type: object
                  properties:
                    name: { $ref: '#/components/schemas/Translations', readOnly: true }
            cancelled: { type: boolean }
            release_date_precision: { type: string }
...
    Statistic_details:
      type: object
      properties:
        version: { type: integer }
        shortname: { type: string, readOnly: true }
        approval_status: { type: string, nullable: true }
        main_language: { type: string }
        division:
          type: object
          properties:
            code: { type: string, nullable: true }
            name: { $ref: '#/components/schemas/Translations', readOnly: true }
        first_released_at: { type: string, format: date, nullable: true }
        yearly_reporting: { type: boolean }
        status:
          type: object
          properties:
            code: { type: string }
        previous_topic_codes: { type: string, nullable: true }
        relation:
          type: object
          properties:
            shortname: { type: string }
            name: { $ref: '#/components/schemas/Translations', readOnly: true }
        name: { $ref: '#/components/schemas/Translations' }
        updated_at: { type: string, format: date-time, readOnly: true }
        comment: { type: string }
        created_at: { type: string, format: date-time, readOnly: true }
        variants:
          type: array
          items: { $ref: '#/components/schemas/Variant' }
        contacts:
          type: array
          items:
            type: object
            properties:
              username: { type: string, nullable: true }
              email: { type: string }
              name: { type: string }
        statistic_region_levels:
          type: array
          items: { $ref: '#/components/schemas/Translations' }
    Statistic_listing:
      type: object
      properties:
        shortname: { type: string, readOnly: true }
        main_language: { type: string }
        status:
          type: object
          properties:
            code: { type: string }
        name: { $ref: '#/components/schemas/Translations' }
        contacts:
          type: array
          items:
            type: object
            properties:
              username: { type: string, nullable: true }
              email: { type: string }
    Translations:
      type: array
      items:
        type: object
        properties:
          language_code: { type: string }
          text: { type: string }
    Variant:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        version: { type: integer }
        updated_at: { type: string, format: date-time, readOnly: true }
        level_of_detail:
          type: object
          properties:
            name: { $ref: '#/components/schemas/Translations', nullable: true }
        created_at: { type: string, format: date-time, readOnly: true }
        cancelled: { type: boolean }
        freqency:
          type: object
          properties:
            name: { $ref: '#/components/schemas/Translations', readOnly: true }
        revision: { type: string }
paths:
  /statistics:
    get:
      summary: List of all statistics
      tags: [statistics]
      responses:
        '200':
          description: List of statistics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Statistic_listing' }
    post:
      summary: Create a new statistic
      tags: [statistics]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Statistic_details' }
      responses:
        '201':
          description: Statistic details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Statistic_details' }
  /statistics/{shortname}:
    get:
      summary: Get a statistic by its shortname
      tags: [statistics]
      parameters:
        - name: shortname
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statistic details with all its variants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Statistic_details' }
    put:
      summary: Update a statistic
      tags: [statistics]
      parameters:
        - name: shortname
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Statistic_details' }
      responses:
        '200':
          description: Updated statistic details with all its variants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Statistic_details' }
...
  /releases/{id}:
    get:
      summary: Get release by id
      tags: [releases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Release details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Release_details' }
    put:
      summary: Update a release
      tags: [releases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Release_update' }
      responses:
        '200':
          description: Updated release details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Release_details' }
...